Modify TDC mask from 32 to 64 bit

+++ CondFormats/HcalObjects/src/HcalTPParameters.cc
@@ -11,7 +11,7 @@ HcalTPParameters::HcalTPParameters() : version_(0), adcCut_(0), tdcMask_(0),
 
 HcalTPParameters::~HcalTPParameters() { }
 
-void HcalTPParameters::loadObject(int version, int adcCut, uint32_t tdcMask, 
+void HcalTPParameters::loadObject(int version, int adcCut, uint64_t tdcMask, 
 				  uint32_t tbits, int auxi1,  int auxi2) {
   version_ = version;
   adcCut_  = adcCut;
